debugged

+++ controllers/userController.js
@@ -1,4 +1,5 @@
 const { User, Thought } = require("../models");
+const { Types } = require("mongoose");
 
 // GET all users
 const getUsers = async (req, res) => {
@@ -22,6 +23,7 @@ const getSingleUser = async (req, res) => {
     }
     res.status(200).json(user);
   } catch (error) {
+    console.log(error);
     res.status(500).json(error);
   }
 };
@@ -32,7 +34,13 @@ const createUser = async (req, res) => {
     const newUser = await User.create(req.body);
     res.status(200).json(newUser);
   } catch (error) {
-    res.status(500).json(error);
+    if (error.code === 11000) {
+      // Duplicate key error
+      res.status(400).json({ message: "Username or email already exists" });
+    } else {
+      console.log(error);
+      res.status(500).json(error);
+    }
   }
 };
 
@@ -57,6 +65,11 @@ const updateUser = async (req, res) => {
 // DELETE user by id
 const deleteUser = async (req, res) => {
   try {
+    // check if userId is a valid id in user database
+    if (!Types.ObjectId.isValid(req.params.userId)) {
+      return res.status(400).json({ message: "No user with this ID!" });
+    }
+
     const user = await User.findByIdAndDelete(req.params.userId);
     if (!user) {
       return res.status(404).json({ message: "No user with this ID!" });
@@ -69,6 +82,7 @@ const deleteUser = async (req, res) => {
       message: "User and associated thoughts have been successfully deleted!",
     });
   } catch (error) {
+    console.log(error);
     res.status(500).json(error);
   }
 };
@@ -76,13 +90,36 @@ const deleteUser = async (req, res) => {
 // POST to add a new friend
 const addFriend = async (req, res) => {
   try {
-    const friend = await User.findById(req.params.friendId);
-    if (!friend) {
-      return res.status(404).json({ message: "Friend not found!" });
+    console.log(
+      "userId: ",
+      req.params.userId,
+      "freindId: ",
+      req.params.friendId
+    );
+    // check if friendId is a valid id in user database
+    if (!Types.ObjectId.isValid(req.params.friendId)) {
+      return res
+        .status(400)
+        .json({ message: "Cannot add friend: No user with this ID!" });
+    }
+    // check if userId is a valid id in user database
+    if (!Types.ObjectId.isValid(req.params.userId)) {
+      return res.status(400).json({ message: "No user with this ID!" });
     }
 
-    const user = await User.findById(req.params.userId);
+    const user = await User.findByIdAndUpdate(
+      req.params.userId,
+      {
+        addToSet: { friends: req.params.friendId },
+      },
+      { new: true }
+    );
     if (!user) {
+      return res.status(404).json({ message: "No user with this ID!" });
+    }
+
+    const friend = await User.findById(req.params.friendId);
+    if (!friend) {
       return res.status(404).json({ message: "No user with this ID!" });
     }
 
@@ -100,15 +137,22 @@ const addFriend = async (req, res) => {
       message: `${friend.username} has been added to your friend list!`,
     });
   } catch (error) {
+    console.log(error);
     res.status(500).json(error);
   }
 };
 
 const deleteFriend = async (req, res) => {
   try {
-    const friend = await User.findById(req.params.friendId);
-    if (!friend) {
-      return res.status(404).json({ message: "Friend not found!" });
+    // check if friendId is a valid id in user database
+    if (!Types.ObjectId.isValid(req.params.friendId)) {
+      return res
+        .status(400)
+        .json({ message: "Cannot delete friend: No user with this ID!" });
+    }
+    // check if userId is a valid id in user database
+    if (!Types.ObjectId.isValid(req.params.userId)) {
+      return res.status(400).json({ message: "No user with this ID!" });
     }
 
     const user = await User.findByIdAndUpdate(
@@ -116,14 +160,28 @@ const deleteFriend = async (req, res) => {
       { $pull: { friends: req.params.friendId } },
       { new: true }
     );
-
     if (!user) {
       return res.status(404).json({ message: "No user with this ID!" });
     }
+
+    const friend = await User.findById(req.params.friendId);
+    if (!friend) {
+      return res.status(404).json({ message: "No user with this ID!" });
+    }
+
+    // check if friendId exists in user friend list
+    const friended = user.friends.includes(req.params.friendId);
+    if (!friended) {
+      return res.status(400).json({
+        message: "Cannot delete friend: This user is not in your friend list",
+      });
+    }
+
     res.status(200).json({
       message: `${friend.username} has been deleted from your friend list!`,
     });
   } catch (error) {
+    console.log(error);
     res.status(500).json(error);
   }
 };
